Modularized the game models into separate files, replaced monolithic `models.py` file with `models` directory
